[FIX] stock_picking_report_custom_description: avoid dupe product name
By default the sale description is set as
f"{product_name}\n{description". If we pass that string as the
description_picking, when we print the report, the name will be
duplicated. To avoid that effect, we want to trim the product name from
the description passed to the picking by default.

TT52044