[IMP] stock_picking_report_custom_description: update description
Before this change only when the procurements where updated the sale
line description would be propagated so if a user updated that
description after the sale was confirmed it wouldn't be written to the
picking move.

Now every time the user changes the line it will be synced to the
related moves.

TT52171